Fixes (#70)
Reviewed a bit RenderToCube test.
Fence timeout is now an uint64_t (as should always have been).
Reworked memory barriers, for buffers and images.
Fixed a compilation bug in Mat4.hpp
Fixed a bug in GlPipeline, that produced an assertion in Signal.
Reworked debug logs.

+++ source/Renderer/GlRenderer/Src/Core/GlRenderingResources.hpp
@@ -37,7 +37,7 @@ namespace gl_renderer
 		*\return
 		*	\p true si l'attente n'est pas sortie en timeout.
 		*/
-		bool waitRecord( uint32_t timeout )override;
+		bool waitRecord( uint64_t timeout )override;
 	};
 }
